requirements.txt
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.86.0
torch==2.0.1
torchvision==0.15.2
transformers==4.30.2
moviepy==1.0.3
numpy==1.24.3
pytest==7.3.1
open_clip_torch==2.20.0
laion_clap==1.1.0
qdrant-client==1.6.0
langchain==0.0.350
langchain-community==0.0.10
sentence-transformers==2.2.2
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
ollama==0.1.4
huggingface-hub==0.25.2